[feature 162469296] Make changes to the patch admin, and to admin validations

+++ app/api/v2/views/redflag_views.py
@@ -6,6 +6,7 @@ from app.api.v2.models.redflag_models import IncidentsModel
 from app.api.v2.models.user_models import UsersModel
 import smtplib
 
+
 UsersModel = UsersModel()
 
 
@@ -22,7 +23,7 @@ class MyIncidents(Resource, IncidentsModel):
         data = request.get_json(force=True)
         try:
             if not data:
-                return {"message": "Kindly input user info"}, 200
+                return {"message": "Kindly input user info"}, 400
             elif not data['images'] or not data['comment']:
                 return {"message":
                         "Ensure you've filled all field. i.e {}".format(data)}, 400
@@ -90,20 +91,14 @@ class MyRecords(Resource, IncidentsModel):
     @jwt_required
     def get(self, id):
         """ Get a specific red-flag record """
-        username = UsersModel.get_username_user(username)
-        print(username)
-        current_user = get_jwt_identity()
-        print(current_user)
-        if username == current_user:
-            incidents = self.db.get_incident_by_id(id)
-            if incidents:
+        incidents = self.db.get_incident_by_id(id)
+        if incidents:
 
-                    return make_response(jsonify({
-                        "data": incidents
-                    }), 200)
+                return make_response(jsonify({
+                    "data": incidents
+                }), 200)
 
-            return {"error": "Incident with that id not found"}, 404
-        return {"error": "You aren't authorized to access this route"}, 403
+        return {"error": "Incident with that id not found"}, 404
 
     @jwt_required
     def delete(self, id):
@@ -111,12 +106,15 @@ class MyRecords(Resource, IncidentsModel):
         incidents = self.db.get_incident_by_id(id)
         if not incidents:
             return {"error": "Incident with that id not found"}, 404
-        current_user = get_jwt_identity()
-        print(current_user)
-        deleting = self.db.delete_redflag(id)
-
-        if deleting:
-            return {"id": id, "message": deleting}, 200
+        else:
+            created_by = self.db.get_created_by(id)
+            current_user = get_jwt_identity()
+            print(current_user)
+            print(created_by)
+            if created_by != current_user:
+                delete = self.db.delete_redflag(id)
+                return {"message": delete}, 200
+            return {"message": "You are not allowed to perform this action"}, 403
 
 
 class MySpecificRecords(Resource, IncidentsModel):
@@ -150,10 +148,6 @@ class MyCommentRecords(Resource, IncidentsModel):
     def patch(self, id):
         """ Allows you to make changes to redflag commments"""
         data = request.get_json(force=True)
-        # comment = data['comment']
-        # check_comment = self.db.check_existing_comment(comment)
-        # if check_comment:
-        #     return {"Message": "Comment already exists"}
         get_by_id = self.db.get_incident_by_id(id=id)
 
         if get_by_id:
@@ -174,12 +168,12 @@ class MyStatusRecords(Resource, IncidentsModel):
         """ Admin makes changes to status"""
         fetch_by_id = self.db.get_incident_by_id(id=id)
         data = request.get_json(force=True)
-        if data['status'] != 'draft':
-            return {
-                "message": "Forbidden, status can only be updated when draft"
-            }, 403
-        elif fetch_by_id:
-            self.db.update_status(data['status'], id)
-            return {"New status": data['status'],
-                    "message": "Updated status successfully"}, 200
-        return {"message": " Status Incident not found"}, 404
+        current_user = get_jwt_identity()
+        is_admin = self.db.get_user_role(current_user)
+        if is_admin:
+            if fetch_by_id:
+                self.db.update_status(data['status'], id)
+                return {"New status": data['status'],
+                        "message": "Updated status successfully"}, 200
+            return {"message": " Status Incident not found"}, 404
+        return {"message": "You are not allowed to perform this action"}, 403
